refactor : StringUtil 클래스 생성
 - LottoNumber 내부의 isBlank와 isNumber를 Util 클래스로 분리
 - isBlank 내부에서 null 체크를 먼저 하게해서 nullException 일어나지 않도록 변경
